+++ src/dist_s1/__main__.py
@@ -100,6 +100,13 @@ def common_options(func: Callable) -> Callable:
         help='S3 bucket to upload the final products to.',
     )
     @click.option(
+        '--n_workers_for_despeckling',
+        type=int,
+        default=5,
+        required=False,
+        help='N CPUs to use for despeckling the bursts',
+    )
+    @click.option(
         '--bucket_prefix',
         type=str,
         default='',
@@ -141,7 +148,8 @@ def run_sas_prep(
     product_dst_dir: str | Path | None,
     bucket: str | None,
     bucket_prefix: str,
-) -> str:
+    n_workers_for_despeckling: int,
+) -> None:
     """Run SAS prep workflow."""
     run_config = run_dist_s1_sas_prep_workflow(
         mgrs_tile_id,
@@ -160,6 +168,7 @@ def run_sas_prep(
         product_dst_dir=product_dst_dir,
         bucket=bucket,
         bucket_prefix=bucket_prefix,
+        n_workers_for_despeckling=n_workers_for_despeckling,
     )
     run_config.to_yaml(runconfig_path)
 
@@ -167,9 +176,9 @@ def run_sas_prep(
 # SAS Workflow (No Internet Access)
 @cli.command(name='run_sas')
 @click.option('--runconfig_yml_path', required=True, help='Path to YAML runconfig file', type=click.Path(exists=True))
-def run_sas(runconfig_yml_path: str | Path) -> str:
-    runconfig_data = RunConfigData.from_yaml(runconfig_yml_path)
-    run_dist_s1_sas_workflow(runconfig_data)
+def run_sas(runconfig_yml_path: str | Path) -> None:
+    run_config = RunConfigData.from_yaml(runconfig_yml_path)
+    run_dist_s1_sas_workflow(run_config)
 
 
 # Effectively runs the two workflows above in sequence
@@ -192,6 +201,7 @@ def run(
     product_dst_dir: str | Path | None,
     bucket: str | None,
     bucket_prefix: str,
+    n_workers_for_despeckling: int,
 ) -> str:
     """Localize data and run dist_s1_workflow."""
     # Localize data
@@ -212,6 +222,7 @@ def run(
         product_dst_dir=product_dst_dir,
         bucket=bucket,
         bucket_prefix=bucket_prefix,
+        n_workers_for_despeckling=n_workers_for_despeckling,
     )
     return run_config