+++ src/dist_s1/workflows.py
@@ -116,7 +116,7 @@ def run_despeckle_workflow(run_config: RunConfigData) -> None:
     rtc_paths = copol_paths + crosspol_paths
     dst_paths = dspkl_copol_paths + dspkl_crosspol_paths
 
-    despeckle_and_serialize_rtc_s1(rtc_paths, dst_paths)
+    despeckle_and_serialize_rtc_s1(rtc_paths, dst_paths, n_workers=run_config.n_workers_for_despeckling)
 
 
 def run_normal_param_estimation_workflow(run_config: RunConfigData) -> None:
@@ -312,6 +312,7 @@ def run_dist_s1_sas_prep_workflow(
     product_dst_dir: str | Path | None = None,
     bucket: str | None = None,
     bucket_prefix: str = '',
+    n_workers_for_despeckling: int = 5,
 ) -> RunConfigData:
     run_config = run_dist_s1_localization_workflow(
         mgrs_tile_id,
@@ -331,6 +332,7 @@ def run_dist_s1_sas_prep_workflow(
     run_config.product_dst_dir = product_dst_dir
     run_config.bucket = bucket
     run_config.bucket_prefix = bucket_prefix
+    run_config.n_workers_for_despeckling = n_workers_for_despeckling
     return run_config
 
 
@@ -361,6 +363,7 @@ def run_dist_s1_workflow(
     product_dst_dir: str | Path | None = None,
     bucket: str | None = None,
     bucket_prefix: str = '',
+    n_workers_for_despeckling: int = 5,
 ) -> Path:
     run_config = run_dist_s1_sas_prep_workflow(
         mgrs_tile_id,
@@ -379,6 +382,7 @@ def run_dist_s1_workflow(
         product_dst_dir=product_dst_dir,
         bucket=bucket,
         bucket_prefix=bucket_prefix,
+        n_workers_for_despeckling=n_workers_for_despeckling,
     )
     _ = run_dist_s1_sas_workflow(run_config)